[SPARK-50090] Refactor ResolveBinaryArithmetic to separate single-node transformation
### What changes were proposed in this pull request?

Refactor ResolveBinaryArithmetic to separate single-node transformation.

### Why are the changes needed?

Single-pass resolver utilizes these single-node transformations and separations is needed between Catalyst and single-pass logic.

### Does this PR introduce _any_ user-facing change?

### How was this patch tested?

Existing tests

### Was this patch authored or co-authored using generative AI tooling?

Closes #48623 from mihailotim-db/mihailotim-db/refactor_binary_arith.

Authored-by: Mihailo Timotic <mihailo.timotic@databricks.com>
